requirements.txt
pandas
numpy
scikit-learn
torch
pyyaml
mplhep